Convert int to Uint32?

+++ PlaybackSpeedTest_ModifiedDelay2/ISRs_B.c
@@ -132,8 +132,8 @@ interrupt void Codec_ISR()
 		roundedPlaybackIndex = round(playbackIndex);
 	}
 	
-	yLeft = buffer[LEFT][roundedPlaybackIndex];
-	yRight = buffer[RIGHT][roundedPlaybackIndex];
+	yLeft = buffer[LEFT][(Uint32) roundedPlaybackIndex];
+	yRight = buffer[RIGHT][(Uint32) roundedPlaybackIndex];
 	
 
 	CodecDataOut.Channel[LEFT] = yLeft;   // setup the LEFT value
